fixed user & token services && create refresh() & checkAuthState()

+++ apps/backend/src/app/user/services/token/token.service.ts
@@ -4,6 +4,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Token, TokenDocument } from '../../schemas/token.schema';
 
+export type TokenType = 'ACCESS_TOKEN' | 'REFRESH_TOKEN';
+
 export class TokenService {
   constructor(
     @InjectModel(Token.name) private tokenModel: Model<TokenDocument>,
@@ -11,7 +13,7 @@ export class TokenService {
   ) {
   }
 
-  generateTokens(payload): {accessToken: string, refreshToken: string} {
+  public generateTokens(payload): {accessToken: string, refreshToken: string} {
     const accessToken = jwt.sign(payload, this.configService.get('JWT_ACCESS_SECRET'), {expiresIn: '30m'});
     const refreshToken = jwt.sign(payload, this.configService.get('JWT_REFRESH_SECRET'), {expiresIn: '30d'});
 
@@ -20,7 +22,7 @@ export class TokenService {
       refreshToken,
     }
   }
-  async saveToken(userId, refreshToken) {
+  public async saveToken(userId, refreshToken) {
     const tokenData = await this.tokenModel.findOne({user: userId})
     if (tokenData) {
       tokenData.refreshToken = refreshToken;
@@ -29,8 +31,26 @@ export class TokenService {
     return await this.tokenModel.create({ user: userId, refreshToken });
   }
 
-  async removeAll() {
-    await this.tokenModel.deleteMany();
+  async removeToken(refreshToken: string) {
+    await this.tokenModel.deleteOne({refreshToken});
     return null;
+  }
+
+  async findToken(refreshToken: string): Promise<TokenDocument | null> {
+    const token = await this.tokenModel.findOne({refreshToken});
+    return token ? token : null;
+  }
+
+  async removeAll(): Promise<null> {
+    if (this.configService.get('MODE') === 'DEV') {
+      await this.tokenModel.deleteMany();
+      return null;
+    }
+  }
+
+  validateToken(type: TokenType, token: string): TokenDocument {
+    console.log('validateToken', token);
+    const jwtSecret = type === 'ACCESS_TOKEN' ? 'JWT_ACCESS_SECRET' : 'JWT_REFRESH_SECRET'
+    return jwt.verify(token, this.configService.get(jwtSecret))
   }
 }
